Move Builder class inner RuntimeEnv
